refactor waste management operation runtime
